+++ tests/util/evaluate_graph.cpp
@@ -30,7 +30,7 @@ std::vector<torch::jit::IValue> EvaluateGraph(const torch::jit::Block* b, std::v
     if (eval) {
       if (eval.value().isTuple()) {
         auto eval_list = eval.value().toTuple();
-        for (int i = 0; i < eval_list->elements().size(); i++) {
+        for (size_t i = 0; i < eval_list->elements().size(); i++) {
           auto eval_output = eval_list.get()->elements()[i];
           LOG_DEBUG(
               ctx->logger,
